Server_Cam/requirements.txt
picamera
cv2
RPi.GPIO
Flask